Generate associated types from more than just object type shapes
This allows associated types to be generated from enumerable shapes, which were missing before.
