+++ src/PolyType/SourceGenModel/SourceGenTypeShape.cs
@@ -1,5 +1,6 @@
 ﻿using PolyType.Abstractions;
 using System.Reflection;
+using System.Text;
 
 namespace PolyType.SourceGenModel;
 
@@ -23,6 +24,11 @@ public abstract class SourceGenTypeShape<T> : ITypeShape<T>
     ICustomAttributeProvider? ITypeShape.AttributeProvider => typeof(T);
 
     /// <summary>
+    /// Gets the shape of an associated type, by its name.
+    /// </summary>
+    public Func<string, ITypeShape?>? AssociatedTypeShapes { get; init; }
+
+    /// <summary>
     /// Accepts an <see cref="TypeShapeVisitor"/> for strongly-typed traversal.
     /// </summary>
     /// <param name="visitor">The visitor to accept.</param>
@@ -34,5 +40,64 @@ public abstract class SourceGenTypeShape<T> : ITypeShape<T>
     object? ITypeShape.Invoke(ITypeShapeFunc func, object? state) => func.Invoke(this, state);
 
     /// <inheritdoc/>
-    public virtual ITypeShape? GetAssociatedTypeShape(Type associatedType) => null;
+    public ITypeShape? GetAssociatedTypeShape(Type associatedType)
+    {
+        if (associatedType.IsGenericTypeDefinition && typeof(T).GenericTypeArguments.Length != associatedType.GetTypeInfo().GenericTypeParameters.Length)
+        {
+            throw new ArgumentException("Type is not a generic type definition or does not have an equal count of generic type parameters with this type shape.");
+        }
+
+        StringBuilder builder = new();
+        ConstructStableName(associatedType, builder);
+        return AssociatedTypeShapes?.Invoke(builder.ToString());
+    }
+
+    private static void ConstructStableName(Type type, StringBuilder builder)
+    {
+        // The string created here must match the string created in the source generator (Parser.CreateAssociatedTypeId).
+        if (type.DeclaringType is not null)
+        {
+            ConstructStableName(type.DeclaringType, builder);
+            builder.Append('+');
+        }
+        else if (!string.IsNullOrEmpty(type.Namespace))
+        {
+            builder.Append(type.Namespace);
+            builder.Append('.');
+        }
+
+        if (type.IsGenericType)
+        {
+            string nameNoArity = type.Name[..type.Name.IndexOf('`')];
+            builder.Append(nameNoArity);
+            builder.Append('<');
+            if (type.IsGenericTypeDefinition)
+            {
+                builder.Append(',', GetOwnGenericTypeParameterCount(type) - 1);
+            }
+            else
+            {
+                for (int i = 0; i < type.GenericTypeArguments.Length; i++)
+                {
+                    if (i > 0)
+                    {
+                        builder.Append(',');
+                    }
+
+                    ConstructStableName(type.GenericTypeArguments[i], builder);
+                }
+            }
+
+            builder.Append('>');
+        }
+        else
+        {
+            builder.Append(type.Name);
+        }
+
+        static int GetOwnGenericTypeParameterCount(Type type)
+            => type.DeclaringType?.IsGenericType is true
+                ? type.GetTypeInfo().GenericTypeParameters.Length - type.DeclaringType.GetTypeInfo().GenericTypeParameters.Length
+                : type.GetTypeInfo().GenericTypeParameters.Length;
+    }
 }
